Use uint32_t for xalloc

+++ src/xalloc.h
@@ -1,8 +1,8 @@
 #ifndef __XALLOC_H__
 #define __XALLOC_H__
 #include <stdint.h>
-void *xalloc(size_t size);
-void *xalloc0(size_t size);
+void *xalloc(uint32_t size);
+void *xalloc0(uint32_t size);
 void xfree(void *ptr);
-void *xrealloc(void *ptr, size_t size);
+void *xrealloc(void *ptr, uint32_t size);
 #endif /* __XALLOC_H__ */
